feat(/mail): add feature with EJS for email template
