+++ mail/src/routes/sendMailResetPassword.route.ts
@@ -1,9 +1,8 @@
 import express from 'express';
-//import { sendResetPassword } from '../modules/users/controller';
-import { activateLogin } from '../modules/users/controller';
+import { sendResetPassword } from '../modules/users/controller';
 const router = express.Router();
 
 // ⚙️ route to process email reset
-router.post('/mail/login/init', activateLogin);
+router.post('/mail/login/init', sendResetPassword);
 
 export default router;
