+++ mail/src/utils/mails/send.mail.ts
@@ -1,0 +1,26 @@
+import path from 'path';
+import { sendEmail } from '../transporter';
+import ejs from 'ejs';
+
+type EmailOptions = {
+  email: string;
+  subject: string;
+  emailTemplate: string; // 💡 represents file that represents the body of the email to be sent (.ejs file)
+  dataTemplate: { url: string }; // 💡 object that represents the values ​​to include in the body of the email (for exemple, username or any other value)
+};
+export const sendEmailByTemplate = async ({
+  email,
+  subject,
+  emailTemplate,
+  dataTemplate,
+}: EmailOptions): Promise<void> => {
+  const emailHtml = await ejs.renderFile(__dirname + `/../../views/templates/${emailTemplate}`, {
+    dataTemplate,
+  });
+
+  await sendEmail({
+    to: email,
+    subject: subject,
+    html: emailHtml,
+  });
+};
