--- mail/src/views/templates/activateLogin.ejs
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Hôpital DoctoPlan</title>
-  <%- include('./includes/header.ejs') %>
-</head>
-
-<body>
-  <div class="header">
-    <h1>DoctoPlan</h1>
-    <p>Activation du mail utilisateur</p>
-  </div>
-
-  <div class="content">
-    <h2>Bonjour !</h2>
-    <p>Votre compte à été crée. </p>
-    <p>Cliquez sur le lien ci-dessous pour activer votre compte :</p>
-
-    <div class="mdp">
-      <a href="<%= templateData.url %>" class="button">Cliquez ici pour activer votre compte</a>
-    </div>
-    <p>ou copiez collez le lien ci-dessous dans le votre navigateur</p>
-
-    <a href="<%= templateData.url %>">
-      <%= templateData.url %>
-    </a>
-
-
-    <%- include('./includes/footer.ejs') %>
-</body>
-
-</html>
